+++ tests/TestCase/Utility/ThumbCreatorTest.php
@@ -33,7 +33,6 @@ class ThumbCreatorTest extends TestCase
     public function testConstructNoExistingFile()
     {
         $this->expectException(NotReadableException::class);
-        $this->expectExceptionMessage('File or directory `' . add_slash_term(rtr(WWW_ROOT)) . 'img' . DS . 'noExistingFile.gif` is not readable');
         $this->getThumbCreatorInstance('noExistingFile.gif');
     }
 
@@ -45,7 +44,6 @@ class ThumbCreatorTest extends TestCase
     {
         Plugin::load('TestPlugin');
         $this->expectException(NotReadableException::class);
-        $this->expectExceptionMessage('File or directory `' . add_slash_term(rtr(Plugin::path('TestPlugin'))) . 'webroot' . DS . 'img' . DS . 'noExistingFile.gif` is not readable');
         $this->getThumbCreatorInstance('TestPlugin.noExistingFile.gif');
     }