+++ tests/TestCase/Utility/ThumbManagerTest.php
@@ -92,6 +92,12 @@ class ThumbManagerTest extends TestCase
      */
     public function testGetAll()
     {
-        $this->assertGreaterThanOrEqual(3, $this->ThumbManager->getAll());
+        $result = $this->ThumbManager->getAll();
+        $resultWithSort = $this->ThumbManager->getAll(true);
+        $this->assertCount(3, $result);
+        $this->assertCount(3, $resultWithSort);
+        $this->assertEquals($result, $resultWithSort);
+        $this->skipIf(IS_WIN);
+        $this->assertNotSame($result, $resultWithSort);
     }
 }